+++ cloudmanager/resource_netapp_cloudmanager_cvo_aws.go
@@ -349,7 +349,6 @@ func resourceCVOAWS() *schema.Resource {
 			"retries": {
 				Type:     schema.TypeInt,
 				Optional: true,
-				ForceNew: true,
 				Default:  60,
 			},
 		},