+++ cloudmanager/resource_netapp_cloudmanager_cvo_azure.go
@@ -304,7 +304,6 @@ func resourceCVOAzure() *schema.Resource {
 			"retries": {
 				Type:     schema.TypeInt,
 				Optional: true,
-				ForceNew: true,
 				Default:  60,
 			},
 		},