+++ cloudmanager/resource_netapp_cloudmanager_cvo_gcp.go
@@ -360,7 +360,6 @@ func resourceCVOGCP() *schema.Resource {
 			"retries": {
 				Type:     schema.TypeInt,
 				Optional: true,
-				ForceNew: true,
 				Default:  60,
 			},
 		},